Added base a variety of base conversion functions for 13 different data types, added strlen for wide character strings, and more.

+++ src/ft_numlength.c
@@ -6,13 +6,13 @@
 /*   By: dcelojev <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/01 17:03:10 by dcelojev          #+#    #+#             */
-/*   Updated: 2019/05/01 17:52:18 by dcelojev         ###   ########.fr       */
+/*   Updated: 2019/05/28 16:54:20 by dcelojev         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_numlength(long n)
+size_t	ft_numlength(long long n)
 {
 	size_t length;
